==> 01-cleaning.ipynb <==
# Data Cleaning
Die Rohdaten werden aus dem CSV eingelesen. Fehlende Daten werden entfernt und die Verspätung wird anhand 
von geplanter und tatsächlicher Abfahrtszeit berechnet.

import pandas as pd
import numpy as np

# Load data with right types.
dtypes = {
    'Datum': str,
    'Wochentag': str,
    'Linienverlauf': str,
    'SollAb': str,
    'GPS-Breite Soll': np.float64,
    'GPS-Länge Soll': np.float64,
}

df_linie1_hin_1 = pd.read_csv(filepath_or_buffer='Echtzeitdaten_Linie1_hin_sheet1.csv', 
                              sep=';', dtype=dtypes, thousands=',')
df_linie1_hin_2 = pd.read_csv(filepath_or_buffer='Echtzeitdaten_Linie1_hin_sheet2.csv', 
                              sep=';', dtype=dtypes, thousands=',')
df_linie1_hin_3 = pd.read_csv(filepath_or_buffer='Echtzeitdaten_Linie1_hin_sheet3.csv', 
                              sep=';', dtype=dtypes, thousands=',')
df_linie1_hin_4 = pd.read_csv(filepath_or_buffer='Echtzeitdaten_Linie1_hin_sheet4.csv', 
                              sep=';', dtype=dtypes, thousands=',')
df_linie1_hin_1.head()

# Check that parsing the right data types worked.
df_linie1_hin_1.dtypes

# Combine the data frames into a single frame.
df_linie1_hin = pd.concat([df_linie1_hin_1, df_linie1_hin_2, df_linie1_hin_3, df_linie1_hin_4], 
                          ignore_index=True)
df_linie1_hin.head()

# How many departure times are missing.
len(df_linie1_hin[df_linie1_hin['IstAb'] == ' '])

# Drop rows with missing values.
df_linie1_hin = df_linie1_hin.drop(index=(df_linie1_hin[df_linie1_hin['IstAb'] == ' ']).index)
df_linie1_hin = df_linie1_hin.reset_index(drop=True)
len(df_linie1_hin[df_linie1_hin['IstAb'] == ' '])

# We still have enough values to work with.
df_linie1_hin.shape

# Convert departure times to datetime objects while setting the correct date.
df_linie1_hin.loc[:, 'SollAb'] = pd.to_datetime(df_linie1_hin['Datum'] + ' ' + df_linie1_hin['SollAb'], format='%d.%m.%Y %H:%M:%S')
df_linie1_hin.loc[:, 'IstAb'] = pd.to_datetime(df_linie1_hin['Datum'] + ' ' + df_linie1_hin['IstAb'], format='%d.%m.%Y %H:%M:%S')
df_linie1_hin.loc[:, 'Datum'] = pd.to_datetime(df_linie1_hin['Datum'], format='%d.%m.%Y')
df_linie1_hin.dtypes

# Discard arrival times as they are not really important.
df_linie1_hin = df_linie1_hin.drop(columns=['SollAn', 'IstAn'])
df_linie1_hin.head()

# Calculate the delay.
df_linie1_hin['Verspätung'] = df_linie1_hin['IstAb'] - df_linie1_hin['SollAb']
df_linie1_hin.head()

# Transform delay from timedelta to seconds.
df_linie1_hin['Verspätung_total_seconds'] = df_linie1_hin['Verspätung'].dt.total_seconds()
df_linie1_hin.head()

# Save results.
df_linie1_hin.to_pickle('df_linie1_hin.pickle')
