==> 02_weather_feature.ipynb <==
# 2. Weather feature
Die Wetterdaten aus der Radverkehrszählung der Standmünster werden geladen und in ein passendes Format überführt.

import pandas as pd
import numpy as np
import pickle
import datetime

with open('df_linie1_hin.pickle', 'rb') as fp:
    df = pickle.load(fp)
df.head()

Data can be downloaded from http://www.stadt-muenster.de/verkehrsplanung/verkehr-in-zahlen/radverkehrszaehlungen/wolbecker-strasse.html


# Load bicyle data that contains wheather information.
df_bicyle = pd.read_csv('zaehlstelle_wolbecker_2016_stundenauswertung.csv', skipfooter=2, engine='python')
df_bicyle.head()

# Parse to timestamps.
df_bicyle['Stunden'] = pd.to_datetime(df_bicyle['Stunden'], format='%d/%m/%Y %H:%M')
df_bicyle.dtypes

# Interpolate missing values using the value with forward fill.
df_bicyle['Wetter'] = df_bicyle['Wetter'].fillna(method='ffill')
df_bicyle['Temperatur (°C)'] = df_bicyle['Temperatur (°C)'].fillna(df_bicyle['Temperatur (°C)'].mean())
df_bicyle['Windstärke (km/h)'] = df_bicyle['Windstärke (km/h)'].fillna(df_bicyle['Windstärke (km/h)'].mean())
df_bicyle.head()

# Fake 2017 data, since they are missing.
df_bicyle2017 = df_bicyle.copy()
df_bicyle2017['Stunden'] = df_bicyle2017['Stunden'] + datetime.timedelta(days=366)
df_bicyle2017.head()

# Concat dataframes of 2017 and 2018.
df_bicyle = pd.concat([df_bicyle, df_bicyle2017])
df_bicyle.head()

df_bicyle = df_bicyle.reset_index(drop=True)
df_bicyle.head()

# Drop useless columns.
df_bicyle = df_bicyle.drop(columns=['Wolbecker Straße (gesamt)', 'FR stadteinwärts', 'FR stadtauswärts'])

# Save results.
df_bicyle.to_pickle('df_weather.pickle')
