==> 04_feature_merging.ipynb <==
# 4. Feature Merging
Die erstellenten Features werden zusammengeführt.

import pandas as pd
import pickle
import numpy as np

with open('df_linie1_hin.pickle','rb') as fp:
    df_linie1_hin = pickle.load(fp)

with open('df_holidays.pickle','rb') as fp:
    df_holidays = pickle.load(fp)
    
with open('df_weather.pickle','rb') as fp:
    df_weather = pickle.load(fp)

df_weather.head()

df_linie1_hin_merged = pd.merge(df_linie1_hin, df_holidays, how="left", left_on="Datum", right_on="date" )
df_linie1_hin_merged.head()

#df_linie1_hin_merged[df_linie1_hin_merged["date"]=="2017-08-14"].head()

# Binarize holiday feature.
df_linie1_hin_merged["Feiertag"] = df_linie1_hin_merged["date"].notna()
df_linie1_hin_merged.head()

df_linie1_hin_merged = df_linie1_hin_merged.drop(columns=["holiday_region","name","weekday","date"])
df_linie1_hin_merged.head()

# Sort for merging.
df_linie1_hin_merged = df_linie1_hin_merged.sort_values('SollAb')
df_weather = df_weather.sort_values('Stunden')

# We have weather data for every hour. We want to merge the weather data by nearest timepoint
# to a certain delay.
df_merged = pd.merge_asof(
    left=df_linie1_hin_merged, 
    right=df_weather, 
    left_on='SollAb',
    right_on='Stunden',
    direction='nearest'
)
df_merged.head()

# Check by sample whether the join worked.
df_merged.iloc[1546]

# Check whether the nearest timepoint for the weather data was chosen.
df_weather[df_weather['Stunden'] == '2016-10-10 15:00']

# Serialize output.
df_merged.to_pickle("df_features.pickle")
