==> 05_feature_preprocessing.ipynb <==
# 5. Feature preprocessing
Die Daten werden in eine für *machine learning* nutzbare Represantation überführt.

import pandas as pd
import numpy as np
import pickle
from sklearn import preprocessing

with open('df_features.pickle', 'rb') as fp:
    df_pre = pickle.load(fp)
df_pre.head()

# Discard outliers.
df_pre = df_pre[df_pre["Verspätung_total_seconds"] > -1200]  #discard early departures of more than 20 minutes
df_pre = df_pre[df_pre["Verspätung_total_seconds"] < 3600]   #discard delays longer than an hour
df_pre.head()

# Scaling numerical features to range [0, 1]. This makes training models usually easier.
min_max_scaler = preprocessing.MinMaxScaler()
df_pre.loc[:, ('Temperatur (°C)', 'Windstärke (km/h)')] = min_max_scaler.fit_transform(
    df_pre[['Temperatur (°C)', 'Windstärke (km/h)']]
)

# Endcode categorial variables as one-hot vectors.
df_pre = pd.concat([df_pre, pd.get_dummies(df_pre[['Wochentag', 'Wetter', 'Linienverlauf']])], axis=1)
df_pre.head()

# Extract number of Haltestelle.
df_pre['Haltestelle'] = pd.to_numeric(df_pre['Linienverlauf'].str[:3])

# Add month as label. 
df_pre['month'] = df_pre['Datum'].dt.month
df_pre[['Datum', 'month']].head()

df_pre.head()

df_pre.to_pickle('df_preprocessed.pickle')
